auth_service_alignsys/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.11.1
attrs==25.4.0
bcrypt==4.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==42.0.5
Django==5.0.1
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-spectacular==0.27.0
email-validator==2.3.0
frozenlist==1.8.0
gunicorn==22.0.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
multidict==6.7.1
packaging==26.0
pillow==10.4.0
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==3.0
pydantic_core==2.42.0
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.3
rpds-py==0.30.0
sqlparse==0.5.5
twilio==9.2.3
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.7.0
yarl==1.23.0
